modules/logjam_loader/requirements.txt
functions-framework==3.2.0
betterproto==2.0.0b4
certifi==2023.7.22
environs==6.0.0
python-pachyderm==7.4.0
google-cloud==0.34.0
google-cloud-logging==3.1.2
google-cloud-storage== 2.7.0
grpc-google-iam-v1==0.12.4
grpc-interceptor==0.13.2
grpcio==1.53.0
grpcio-health-checking==1.47.0
grpcio-status==1.47.0
grpcio-tools==1.47.0
grpclib==0.4.2
click==8.1.3
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pandas==2.1.4
pathtools==0.1.2
#psycopg2==2.9.5
psycopg2-binary==2.8.6
pyparsing==3.0.6
setuptools==65.5.1
watchdog==1.0.2
Werkzeug==2.3.8
